+++ docs/navtreedata.js
@@ -56,8 +56,8 @@ var NAVTREE =
 var NAVTREEINDEX =
 [
 "_event_8cpp.html",
-"class_excimontec_1_1_parameters.html#a40764c33d7bc7678a08addd99fdb5753",
-"functions.html"
+"class_excimontec_1_1_parameters.html#a3019332d8bfc7927917a1e87bc4ee77d",
+"dir_68267d1309a1af8e8297ef4c3efbcdba.html"
 ];
 
 var SYNCONMSG = 'click to disable panel synchronisation';